@@ -1,17 +1,23 @@
-# This file and other cmake files in this project where inspired by https://github.com/pabloariasal/modern-cmake-sample
-# Based on the article: https://pabloariasal.github.io/2018/02/19/its-time-to-do-cmake-right/
+# This file and other cmake files in this project where inspired by
+# https://github.com/pabloariasal/modern-cmake-sample Based on the article:
+# https://pabloariasal.github.io/2018/02/19/its-time-to-do-cmake-right/
 
-cmake_minimum_required(VERSION 3.3)
-project(MultiNEAT VERSION 0.7 LANGUAGES C CXX)
+cmake_minimum_required(VERSION 3.15)
+project(
+    MultiNEAT
+    VERSION 0.8
+    LANGUAGES CXX)
 
-set (CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
-set (MultiNEAT_SRC_DIR ${CMAKE_SOURCE_DIR}/src)
+option(ENABLE_TESTING "If tests should be built.")
 
+# Set compiler warnings
 if(MSVC)
     # Force to always compile with W4
     if(CMAKE_CXX_FLAGS MATCHES "/W[0-4]")
-        string(REGEX REPLACE "/W[0-4]" "/W4" CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}")
+        string(REGEX REPLACE "/W[0-4]" "/W4" CMAKE_CXX_FLAGS
+                             "${CMAKE_CXX_FLAGS}")
     else()
         set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /W4")
     endif()
