@@ -30,52 +30,66 @@
 // Description: Declarations for a class dealing with random numbers.
 ///////////////////////////////////////////////////////////////////////////////
 
-#ifdef USE_BOOST_RANDOM
-    #include <boost/random.hpp>
-#else
-    #include <stdlib.h>
-#endif
-
+#include <random>
 #include <vector>
 #include <limits>
+#include <stdexcept>
+#include <cereal/cereal.hpp>
+#include <sstream>
 
 namespace NEAT
 {
 
-class RNG
-{
-    
-#ifdef USE_BOOST_RANDOM
-    boost::random::mt19937 gen;
-#endif
+    class RNG
+    {
+        std::mt19937 rng;
 
-public:
-    // Seeds the random number generator with this value
-    void Seed(long seed);
+    public:
+        RNG();
 
-    // Seeds the random number generator with time
-    void TimeSeed();
+        // Seeds the random number generator with this value
+        void Seed(long seed);
 
-    // Returns randomly either 1 or -1
-    int RandPosNeg();
+        // Seeds the random number generator with time
+        void TimeSeed();
 
-    // Returns a random integer between X and Y
-    int RandInt(int x, int y);
+        // Returns randomly either 1 or -1
+        int RandPosNeg();
 
-    // Returns a random number from a uniform distribution in the range of [0 .. 1]
-    double RandFloat();
+        // Returns a random integer between X and Y
+        int RandInt(int x, int y);
 
-    // Returns a random number from a uniform distribution in the range of [-1 .. 1]
-    double RandFloatSigned();
+        // Returns a random number from a uniform distribution in the range of [0 .. 1]
+        double RandFloat();
 
-    // Returns a random number from a gaussian (normal) distribution in the range of [-1 .. 1]
-    double RandGaussSigned();
+        // Returns a random number from a uniform distribution in the range of [-1 .. 1]
+        double RandFloatSigned();
 
-    // Returns an index given a vector of probabilities
-    int Roulette(std::vector<double>& a_probs);
-};
+        // Returns a random number from a gaussian (normal) distribution in the range of [-1 .. 1]
+        double RandGaussSigned();
 
+        // Returns an index given a vector of probabilities
+        int Roulette(std::vector<double> &a_probs);
 
+        template <class Archive>
+        void save(Archive &ar) const
+        {
+            std::stringstream ss;
+            ss << rng;
+            std::string serialized = ss.str();
+            ar &serialized;
+        }
+
+        template <class Archive>
+        void load(Archive &ar)
+        {
+            std::string serialized;
+            ar &serialized;
+            std::stringstream ss;
+            ss << serialized;
+            ss >> rng;
+        }
+    };
 
 } // namespace NEAT
 
