@@ -1,14 +1,16 @@
 //
 // Created by matteo on 6/7/20.
+// Edited by aart in 2022
 //
 
 #include "../Innovation.h"
-#include <boost/python.hpp>
+#include <pybind11/pybind11.h>
 
-void export_Innovation()
-{
-    using namespace NEAT;
-    boost::python::class_<Innovation>("Innovation", boost::python::init<int, InnovationType, int, int, NeuronType, int>())
+using namespace NEAT;
+
+void export_Innovation(pybind11::module_& mod) {
+    pybind11::class_<Innovation>(mod, "Innovation")
+        .def(pybind11::init<int, InnovationType, int, int, NeuronType, int>())
             .def("ID", &Innovation::ID)
             .def("InnovType", &Innovation::InnovType)
             .def("FromNeuronID", &Innovation::FromNeuronID)
@@ -16,5 +18,4 @@ void export_Innovation()
             .def("NeuronID", &Innovation::NeuronID)
             .def("GetNeuronType", &Innovation::GetNeuronType)
             ;
-
 }
