@@ -1,21 +1,22 @@
 //
 // Created by matteo on 6/7/20.
+// Edited by aart in 2022
 //
 
 #include "../Random.h"
-#include <boost/python.hpp>
+#include <pybind11/pybind11.h>
 
-void export_RNG()
-{
-    boost::python::class_<NEAT::RNG>("RNG", boost::python::init<>())
-            .def("Seed", &NEAT::RNG::Seed)
-            .def("TimeSeed", &NEAT::RNG::TimeSeed)
-            .def("RandPosNeg", &NEAT::RNG::RandPosNeg)
-            .def("RandInt", &NEAT::RNG::RandInt)
-            .def("RandFloat", &NEAT::RNG::RandFloat)
-            .def("RandFloatSigned", &NEAT::RNG::RandFloatSigned)
-            .def("RandGaussSigned", &NEAT::RNG::RandGaussSigned)
-            .def("Roulette", &NEAT::RNG::Roulette)
-            ;
+using namespace NEAT;
 
+void export_RNG(pybind11::module_& mod) {
+    pybind11::class_<RNG>(mod, "RNG")
+        .def(pybind11::init<>())
+        .def("Seed", &RNG::Seed)
+        .def("TimeSeed", &RNG::TimeSeed)
+        .def("RandPosNeg", &RNG::RandPosNeg)
+        .def("RandInt", &RNG::RandInt)
+        .def("RandFloat", &RNG::RandFloat)
+        .def("RandFloatSigned", &RNG::RandFloatSigned)
+        .def("RandGaussSigned", &RNG::RandGaussSigned)
+        .def("Roulette", &RNG::Roulette);
 }
