@@ -1,15 +1,15 @@
 //
 // Created by matteo on 6/7/20.
+// Edited by aart in 2022
 //
 
 #include "../Utils.h"
-#include <boost/python.hpp>
+#include <pybind11/pybind11.h>
 
-void export_SearchMode()
-{
-    boost::python::enum_<SearchMode>("SearchMode")
-            .value("COMPLEXIFYING", COMPLEXIFYING)
-            .value("SIMPLIFYING", SIMPLIFYING)
-            .value("BLENDED", BLENDED)
-            ;
+void export_SearchMode(pybind11::module_& mod) {
+    pybind11::enum_<SearchMode>(mod, "SearchMode")
+        .value("COMPLEXIFYING", COMPLEXIFYING)
+        .value("SIMPLIFYING", SIMPLIFYING)
+        .value("BLENDED", BLENDED)
+        .export_values();
 }
