@@ -3,8 +3,6 @@
 //
 
 #include <iostream>
-#include <boost/archive/text_iarchive.hpp>
-#include <boost/archive/text_oarchive.hpp>
 #include <Genome.h>
 
 
@@ -13,7 +11,7 @@ NEAT::Genome deserialize(const std::string &filename)
     NEAT::Genome g;
     std::ifstream input_file(filename);
     {
-        boost::archive::text_iarchive archive(input_file);
+        cereal::JSONInputArchive archive(input_data);
         archive >> g;
     }
     return g;
